usuniecie klasy publicznej i przeniesienie jej
