+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -45,3 +45,22 @@ public class IntLinkedList {
 		return last;
 	}
 }
+
+class Node {
+
+	protected final int value;
+	private  Node prev;
+	protected Node next;
+	
+	protected Node(int i) {
+		value = i;
+	}
+	
+	protected Node getPrev() {
+		return prev;
+	}
+	
+	protected void setPrev(Node prev) {
+		this.prev = prev;
+	}	
+}
