--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,20 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-public class Node {
-
-	protected final int value;
-	private  Node prev;
-	protected Node next;
-	
-	protected Node(int i) {
-		value = i;
-	}
-	
-	protected Node getPrev() {
-		return prev;
-	}
-	
-	protected void setPrev(Node prev) {
-		this.prev = prev;
-	}	
-}
